remove unused page, fix profile links
