--- src/routes/[page].svelte
@@ -1,9 +1,0 @@
-<script>
-	import Index from './_Index.svelte';
-</script>
-
-<svelte:head>
-	<title>Conduit</title>
-</svelte:head>
-
-<Index />
